requirements.txt
streamlit
matplotlib
pandas
numpy
Pillow
opencv-python